# Fix severe performance regression due to change of table formatting