--- a/src/formatters/table.rs
+++ b/src/formatters/table.rs
@@ -322,30 +322,6 @@ fn should_expand(table_constructor: &TableConstructor) -> bool {
     }
 }
 
-/// A fail-fast check to determine whether the formatted fields are going over the budget.
-/// We format each field one at a time, then add its width to the shape, and check to see if the shape is over budget.
-/// Originally, we used `format_singleline_table()` to check this, which had exponential time complexity for nested tables.
-/// We don't need to format the whole table to see if we are going over budget.
-fn check_table_over_budget<'ast>(
-    ctx: &Context,
-    fields: &Punctuated<'ast, Field<'ast>>,
-    shape: Shape,
-) -> bool {
-    // Use an infinite width shape to force everything onto a single line as much as possible
-    // + 2 = opening brace plus space
-    let mut shape = shape.with_infinite_width() + 2;
-
-    for field in fields {
-        let formatted_field = format_field(ctx, field, TableType::SingleLine, shape).0;
-        shape = shape + (formatted_field.to_string().len() + 2); // 2 = ", "
-        if shape.over_budget() {
-            return true;
-        }
-    }
-
-    false
-}
-
 pub fn format_table_constructor<'ast>(
     ctx: &Context,
     table_constructor: &TableConstructor<'ast>,