Fixes #205, which gives an explanation of what happened.

Our change in 0.9.0, as highlighted in #205, led to an exponential time complexity in formatting deeply nested tables. This is because we attempt to format the table on a single line to see if it is over width. This means, if the table contains more tables, we attempt this *again* for each table field, and this adds up exponentially the deeper the table.

Two solutions were attempted here:
1)  Revert back to our original, where we check the position between the opening and closing brace in the input. This is the cheapest solution (taking O(1) time), but isn't great since we rely on the input: if the input is badly formatted, it will lead to the table being pushed multiline when it could've been collapsed.
2) Implement a fail fast version of what we attempted in 0.9.0. Instead of formatting the whole table on a single line, then checking its width to determine if its over budget, we format a single field at a time and see if it has gone over the column width - if it has, bail early and expand multiline.

This PR adds a `nested_tables.lua` benchmark. Benchmarking this on current trunk took **3.0536s** for this single file.
Solution 1) leads to a performance improvement of **98.5%**, going from 3.0536s -> **44.26ms**
Solution 2) leads to a performance improvement of **40.0118%**, going from 3.0536s -> **1.8318s**. This however, is a performance **regression** of **4038.1%** compared to solution 1 (i.e. v0.8.2). Solution 2 is clearly still exponential (which makes sense, if we have tables early on in the fields).

We will revert back to solution 1, as present in v0.8.2. This is the best solution by far, and it was never actually flagged as problematic by any users before 0.9.0