@JohnnyMorganz 2021-06-21
Here are some example flamegraphs to show the performance impact of what we had:
Current trunk (v0.9.2):
![flamegraph-0 9 2-nested_tables](https://user-images.githubusercontent.com/19635171/122769108-2b698980-d29c-11eb-8824-22ca3b9de145.png)
Using brace ranges (Solution 1):
![flamegraph-braces_range-nested_tables](https://user-images.githubusercontent.com/19635171/122769169-39b7a580-d29c-11eb-9754-ad2a4682cf35.png)
Using fail fast (Solution 2):
![flamegraph-fail_fast-nested_tables](https://user-images.githubusercontent.com/19635171/122769205-44723a80-d29c-11eb-8621-5b3fcf6b76f4.png)

As can be seen in 0.9.2 and solution 2, the majority of the time is taken up by `format_table_constructor`, and `format_expression_internal -> format_table_constructor`. This is due to the exponential blowup for formatting fields inside a table, which are themselves table expressions.